Moved switches to their own package inside materials
