+++ src/com/droidquest/levels/RO6.java
@@ -11,6 +11,7 @@ import com.droidquest.decorations.Graphix;
 import com.droidquest.devices.DirectionalSensor;
 import com.droidquest.items.*;
 import com.droidquest.materials.*;
+import com.droidquest.materials.switches.Switch;
 
 import java.awt.*;
