+++ src/com/droidquest/materials/switches/ElevatorSwitch.java
@@ -1,9 +1,10 @@
-package com.droidquest.items;
+package com.droidquest.materials.switches;
 
 import com.droidquest.Room;
 import com.droidquest.decorations.Arrow;
+import com.droidquest.items.Item;
 import com.droidquest.materials.ElevatorOutPortal;
-import com.droidquest.materials.Switch;
+import com.droidquest.materials.Material;
 
 public class ElevatorSwitch extends Switch {
     private int animationState = 0;
@@ -45,7 +46,7 @@ public class ElevatorSwitch extends Switch {
                 animationState++;
                 break;
             case 3:
-                if (ElevatorOutPortal.outRoom == level.rooms.elementAt(11)) {
+                if (ElevatorOutPortal.outRoom == Material.level.rooms.elementAt(11)) {
                     for (int a = 0; a < room.arrows.size(); a++) {
                         Arrow arrow = room.arrows.elementAt(a);
                         arrow.direction = Arrow.DIR_UP;
