+++ src/com/droidquest/materials/switches/MazeLock.java
@@ -1,6 +1,7 @@
-package com.droidquest.items;
+package com.droidquest.materials.switches;
 
-import com.droidquest.materials.Switch;
+import com.droidquest.items.Item;
+import com.droidquest.materials.Material;
 
 
 public class MazeLock extends Switch {
@@ -12,15 +13,15 @@ public class MazeLock extends Switch {
 
     public void TouchedByItem(Item item) {
         if (paintbrush == null) {
-            paintbrush = level.paintbrush;
+            paintbrush = Material.level.paintbrush;
         }
 
         if (!value) {
-            level.paintbrush = null;
+            Material.level.paintbrush = null;
             value = true;
         }
         else {
-            level.paintbrush = paintbrush;
+            Material.level.paintbrush = paintbrush;
             value = false;
         }
     }
