+++ src/com/droidquest/materials/switches/Switch4B.java
@@ -1,7 +1,7 @@
-package com.droidquest.items;
+package com.droidquest.materials.switches;
 
+import com.droidquest.items.Item;
 import com.droidquest.materials.Material;
-import com.droidquest.materials.Switch;
 
 public class Switch4B extends Switch {
     private transient Switch4A sw = null;
@@ -12,8 +12,8 @@ public class Switch4B extends Switch {
 
     public void TouchedByItem(Item item) {
         if (sw == null) {
-            for (int a = 0; a < level.materials.size(); a++) {
-                Material mat = level.materials.elementAt(a);
+            for (int a = 0; a < Material.level.materials.size(); a++) {
+                Material mat = Material.level.materials.elementAt(a);
                 if (mat instanceof Switch4A) {
                     sw = (Switch4A) mat;
                 }
