+++ src/com/droidquest/materials/switches/Switch4D.java
@@ -1,19 +1,19 @@
-package com.droidquest.items;
+package com.droidquest.materials.switches;
 
+import com.droidquest.items.Item;
 import com.droidquest.materials.Material;
-import com.droidquest.materials.Switch;
 
-public class Switch4C extends Switch {
+public class Switch4D extends Switch {
     private transient Switch4A sw = null;
 
-    public Switch4C() {
-        super(Switch.ROT_DOWN);
+    public Switch4D() {
+        super(Switch.ROT_LEFT);
     }
 
     public void TouchedByItem(Item item) {
         if (sw == null) {
-            for (int a = 0; a < level.materials.size(); a++) {
-                Material mat = level.materials.elementAt(a);
+            for (int a = 0; a < Material.level.materials.size(); a++) {
+                Material mat = Material.level.materials.elementAt(a);
                 if (mat instanceof Switch4A) {
                     sw = (Switch4A) mat;
                 }
@@ -27,3 +27,4 @@ public class Switch4C extends Switch {
         }
     }
 }
+
